refactor: convert auth module to monorepo library
