+++ libs/auth/src/auth.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { FirebaseModule } from '@shared/firebase';
+import { UsersModule } from '@shared/users';
+import { AuthService } from './services';
+import { AuthGuard } from './guards';
+
+@Module({
+    imports: [FirebaseModule, UsersModule],
+    providers: [AuthService, AuthGuard],
+    exports: [AuthService, AuthGuard],
+})
+export class AuthModule {}
