+++ libs/auth/src/index.ts
@@ -1,0 +1,3 @@
+export * from './auth.module';
+export * from './guards';
+export * from './services';
