+++ libs/auth/src/services/auth.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { DecodedIdToken, UserRecord } from 'firebase-admin/auth';
-import { FirebaseService } from '@modules/firebase/services';
+import { FirebaseService } from '@shared/firebase';
 
 @Injectable()
 export class AuthService {
